requirements.txt
django===5.0.6
djangorestframework===3.15.2
django-cors-headers===4.4.0
psycopg2===2.9.9
factory-boy===3.3.0